requirements.txt
yfinance
pandas
numpy
requests
python-dotenv
rich
websocket-client
ta
groq
